added shape constraint to donut chart
