+++ src/toolbox/chartLimits.js
@@ -5,6 +5,7 @@
 export const DEFAULT_CHART_LIMITS = {
   BARCHART_MAX_BARS: 20,
   BARCHART_YEAR_MAX_BARS: 20,
+  DONUT_SHAPE_MAX: 30,
   LINEPLOT_LINE_POINT_MIN: 2,
   LINEPLOT_LINE_MAX: 20,
   STACKED_SHAPE_MAX: 200,
